Extracting setup steps for tests
